Modifications to move application to 64-bit address range

+++ mpfs-rpmsg-bm/src/middleware/rpmsg/include/rpmsg_lite.h
@@ -54,11 +54,11 @@ extern "C" {
 #define RL_VERSION "3.1.2" /*!< Current RPMsg Lite version */
 
 /* Shared memory "allocator" parameters */
-#define RL_WORD_SIZE (sizeof(uint32_t))
+#define RL_WORD_SIZE (sizeof(uint64_t))
 #define RL_WORD_ALIGN_UP(a) \
-    (((((uint32_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? ((((uint32_t)a) & (~(RL_WORD_SIZE - 1U))) + 4U) : ((uint32_t)a))
+    (((((uint64_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? ((((uint64_t)a) & (~(RL_WORD_SIZE - 1U))) + 4U) : ((uint64_t)a))
 #define RL_WORD_ALIGN_DOWN(a) \
-    (((((uint32_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? (((uint32_t)a) & (~(RL_WORD_SIZE - 1U))) : ((uint32_t)a))
+    (((((uint64_t)a) & (RL_WORD_SIZE - 1U)) != 0U) ? (((uint64_t)a) & (~(RL_WORD_SIZE - 1U))) : ((uint64_t)a))
 
 /* Definitions for device types , null pointer, etc.*/
 #define RL_SUCCESS    (0)
