+++ mpfs-rpmsg-bm/src/middleware/rpmsg/rpmsg_lite/rpmsg_lite.c
@@ -949,7 +949,7 @@ struct rpmsg_lite_instance *rpmsg_lite_master_init(void *shmem_addr,
      * Since device is RPMSG Remote so we need to manage the
      * shared buffers. Create shared memory pool to handle buffers.
      */
-    rpmsg_lite_dev->sh_mem_base = (char *)RL_WORD_ALIGN_UP((uint32_t)(char *)shmem_addr + (uint32_t)RL_VRING_OVERHEAD);
+    rpmsg_lite_dev->sh_mem_base = (char *)RL_WORD_ALIGN_UP((uint64_t)(char *)shmem_addr + (uint64_t)RL_VRING_OVERHEAD);
     rpmsg_lite_dev->sh_mem_remaining =
         (RL_WORD_ALIGN_DOWN(shmem_length - (uint32_t)RL_VRING_OVERHEAD)) / (uint32_t)RL_BUFFER_SIZE;
     rpmsg_lite_dev->sh_mem_total = rpmsg_lite_dev->sh_mem_remaining;
@@ -965,7 +965,7 @@ struct rpmsg_lite_instance *rpmsg_lite_master_init(void *shmem_addr,
     for (idx = 0U; idx < 2U; idx++)
     {
         ring_info.phy_addr =
-            (void *)(char *)((uint32_t)(char *)shmem_addr + (uint32_t)((idx == 0U) ? (0U) : (VRING_SIZE)));
+            (void *)(char *)((uint64_t)(char *)shmem_addr + (uint64_t)((idx == 0U) ? (0U) : (VRING_SIZE)));
         ring_info.align     = VRING_ALIGN;
         ring_info.num_descs = RL_BUFFER_COUNT;
 
@@ -1151,13 +1151,13 @@ struct rpmsg_lite_instance *rpmsg_lite_remote_init(void *shmem_addr, uint32_t li
     callback[0]                 = rpmsg_lite_tx_callback;
     callback[1]                 = rpmsg_lite_rx_callback;
     rpmsg_lite_dev->vq_ops      = &remote_vq_ops;
-    rpmsg_lite_dev->sh_mem_base = (char *)RL_WORD_ALIGN_UP((uint32_t)(char *)shmem_addr + (uint32_t)RL_VRING_OVERHEAD);
+    rpmsg_lite_dev->sh_mem_base = (char *)RL_WORD_ALIGN_UP((uint64_t)(char *)shmem_addr + (uint64_t)RL_VRING_OVERHEAD);
 
     /* Create virtqueue for each vring. */
     for (idx = 0U; idx < 2U; idx++)
     {
         ring_info.phy_addr =
-            (void *)(char *)((uint32_t)(char *)shmem_addr + (uint32_t)((idx == 0U) ? (0U) : (VRING_SIZE)));
+            (void *)(char *)((uint64_t)(char *)shmem_addr + (uint64_t)((idx == 0U) ? (0U) : (VRING_SIZE)));
         ring_info.align     = VRING_ALIGN;
         ring_info.num_descs = RL_BUFFER_COUNT;
